fix, refactor: перенести логику метода получения рекомендаций в хранилище фильмов, codeStyle
