+++ src/main/java/ru/yandex/practicum/filmorate/dao/user/UserDbStorage.java
@@ -2,26 +2,20 @@ package ru.yandex.practicum.filmorate.dao.user;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.dao.event.EventDbStorage;
-import ru.yandex.practicum.filmorate.dao.film.FilmDbStorage;
 import ru.yandex.practicum.filmorate.exception.AlreadyExistException;
 import ru.yandex.practicum.filmorate.exception.IllegalIdException;
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.model.event.EventOperation;
 import ru.yandex.practicum.filmorate.model.event.EventType;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import static ru.yandex.practicum.filmorate.exception.AlreadyExistException.USER_ALREADY_EXIST_ADVICE;
 import static ru.yandex.practicum.filmorate.exception.AlreadyExistException.USER_ALREADY_EXIST_MESSAGE;
@@ -34,11 +28,9 @@ public class UserDbStorage implements UserStorage {
     private static final String REMOVE_FROM_FRIEND_MESSAGE = "Пользователи успешно удалены из друзей. Их id: ";
 
     private final JdbcTemplate jdbcTemplate;
-    private final FilmDbStorage filmDbStorage;
 
     @Autowired
-    public UserDbStorage(JdbcTemplate jdbcTemplate, FilmDbStorage filmDbStorage) {
-        this.filmDbStorage = filmDbStorage;
+    public UserDbStorage(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
@@ -216,52 +208,6 @@ public class UserDbStorage implements UserStorage {
                     ILLEGAL_OBJECTS_ID_MESSAGE, id, otherId);
             throw new IllegalIdException(ILLEGAL_OBJECTS_ID_MESSAGE, ILLEGAL_OBJECTS_ID_ADVICE);
         }
-    }
-
-    @Override
-    public List<Film> getRecommendations(Long id) {
-        List<Film> films = new ArrayList<>();
-        List<Long> userFilmsId = new ArrayList<>();
-        Long anotherUserId = null;
-        String userFilmsIdString = "";
-        if (checkIfUserExists(id)) {
-            try {
-                userFilmsId = jdbcTemplate.queryForList(
-                        "SELECT film_id " +
-                        "FROM film_likes " +
-                        "WHERE user_id = ?;", Long.class, id);
-
-                userFilmsIdString = userFilmsId.stream()
-                        .map(String::valueOf)
-                        .collect(Collectors.joining(",", " ", " "));
-
-                anotherUserId = jdbcTemplate.queryForObject(
-                        "SELECT user_id " +
-                        "FROM film_likes " +
-                        "WHERE NOT user_id = ? " +
-                          "AND film_id IN (" + userFilmsIdString + ") " +
-                        "GROUP BY user_id " +
-                        "ORDER BY COUNT(film_id) DESC " +
-                        "LIMIT 1;", Long.class, id);
-            } catch (EmptyResultDataAccessException e) {
-                return Collections.emptyList();
-            }
-            films = jdbcTemplate.query(
-                    "SELECT fl.film_id, " +
-                            "f.name, " +
-                            "f.release_date, " +
-                            "f.description, " +
-                            "f.duration, " +
-                            "mp.mpa_rating_id, " +
-                            "mp.name AS mpa_name " +
-                    "FROM FILM_LIKES AS fl " +
-                    "JOIN FILMS AS f ON fl.film_id = f.film_id " +
-                    "JOIN mpa_rating AS mp ON f.mpa_rating_id = mp.mpa_rating_id " +
-                    "WHERE fl.film_id NOT IN (" + userFilmsIdString + ") " +
-                      "AND user_id = ?;", filmDbStorage.getFilmMapper(), anotherUserId);
-            filmDbStorage.setGenreForFilms(films);
-        }
-        return films;
     }
 
     /*-----Вспомогательные методы-----*/
