+++ src/main/java/ru/yandex/practicum/filmorate/dao/user/UserStorage.java
@@ -1,6 +1,5 @@
 package ru.yandex.practicum.filmorate.dao.user;
 
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
@@ -17,6 +16,4 @@ public interface UserStorage {
     User getUserById(Long userId);
 
     void deleteUserById(Long userId);
-
-    List<Film> getRecommendations(Long id);
 }
