+++ src/main/java/ru/yandex/practicum/filmorate/service/user/UserService.java
@@ -1,6 +1,5 @@
 package ru.yandex.practicum.filmorate.service.user;
 
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
@@ -23,6 +22,4 @@ public interface UserService {
     List<User> getAllFriendsList(Long id);
 
     List<User> getCommonFriends(Long id, Long otherId);
-
-    List<Film> getRecommendations(Long id);
 }
