+++ src/main/java/ru/yandex/practicum/filmorate/service/user/UserServiceImpl.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.filmorate.service.user;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.dao.user.UserDbStorage;
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
@@ -60,10 +59,5 @@ public class UserServiceImpl implements UserService {
     @Override
     public List<User> getCommonFriends(Long id, Long otherId) {
         return userStorage.getCommonFriends(id, otherId);
-    }
-
-    @Override
-    public List<Film> getRecommendations(Long id) {
-        return userStorage.getRecommendations(id);
     }
 }
